+++ .gitignore
@@ -1,0 +1,4 @@
+config.py
+.DS_STORE
+*.pyc
+*.log
