--- bing-rewards-master/BingRewards/src/__init__.py
@@ -1,1 +1,0 @@
-
